# src/main/java/org/example/bankcards/dto/card/request/CardDepositRequestDTO.java
package org.example.bankcards.dto.card.request;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class CardDepositRequestDTO {
    @NotNull
    private Long cardId;

    @NotNull
    @DecimalMin(value = "0.01", message = "Deposit amount must be greater than 0")
    private BigDecimal amount;
}